Update patch set 1
Patch Set 1:

I really dislike the idea of maintaining a separate mms_config.xml for d2spr.  But I don't think enabling the option on all CDMA devices is ideal if the problem isn't widespread across carriers.  Multipart SMS is a superior user experience.

Honestly I'm not certain yet what to do with d2spr here.  I don't know how widespread the problem is on Sprint's network.

Didn't there used to be an Mms app setting for it?  That might actually be ideal, so that people affected could tick it on if needed.  Sadly I don't think there's a great way to deal with it transparently.

Patch-set: 1