Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

New idea:

Continue to do multipart as we do, but have a telephony-level option to clamp the PDU header to multiple single-part messages instead of a stream of multipart PDUs.  The network shouldn't choke on that.  The downside would be that messages wouldn't be reassembled on the receiving end, but it's not clear that Sprint's SMS handling allows for that anyways.

I'll put that together and test the next time I'm in Pittsburgh.

Patch-set: 1