Update patch set 1
Patch Set 1:

Well, I'm not aware of this being a problem on other carriers.  It's almost certainly not a problem on AT&T as toppo tested multipart somewhat extensively a few months back.  It's even only an issue in _parts_ of Sprint's network.

I'd probably keep out of common as true multipart is beneficial--message ordering is preserved on the receiver for example.  However, Sprint appears to throw that information away even in the parts of the network where multipart does work, as best I can tell.

Patch-set: 1